% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estatusbar.R
\docType{data}
\name{estatusbar}
\alias{estatusbar}
\title{Statusbar that estimates the remaining time based on the user-registered work progress.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
estatusbar
}
\description{
This is an R port of a Fortran module that was used in the astrophysical codes SPEV and MRGENESIS.
}
\section{Initialization}{

An R6 object is initalized sing \code{estatusbar$new()} with a no arguments. It should be called by the user when the task starts.
}

\section{Adding new entries}{

As the task progresses, the user can call the function \code{add} to register the fraction of the task that has been completed. The routine \code{add} has only one argument, the fraction (between 0 and 1) that is complete at current time.
}

\section{Displaying the status bar}{

At any point the user can call the \code{display} function to draw the status bar in terminal. This function accepts the following arguments:
width = 40, text = NULL, perc = TRUE, eta = TRUE, digits=0
\describe{
\item{width}{The width of the progress bar. The default values is 40 characters}
\item{text}{A custom text string that can be shown next to the progress bar. By default no text is displayed.}
\item{perc}{Percentage display flag, by default set to TRUE. This controls whether the percentage of progress should be printed.}
\item{eta}{Estimated completion time display flag, by default set to TRUE. This controls whether the ETA should be printed.}
\item{digits}{Total number of digits to be used when printing seconds, by default set to 0. This allows the user to show fractions of a second.}
}
}

\examples{

What follows is the typical usage case. The status bar is registered, and then 20 entries are added. Before each add the system sleeps for a random time interval (between 0 and 2 seconds). Afterward an entry is registered and the display is updated.

est <- estatusbar$new()
for (i in 1:20) {
   Sys.sleep(2 * runif(1))
   est$add(i / 20)
   est$display()
}

}
\keyword{datasets}
